refactor: split database.py into db/ package and add unit tests
- Extract 4170-line database.py into focused modules:
  db/config.py (DB_PATH), db/exercises_data.py (INITIAL_EXERCISES),
  db/schema.py (init_db), db/games.py, db/exercises.py, db/users.py
- database.py becomes a one-screen compatibility shim (zero breaking changes)
- Add backend/tests/test_game_engine.py: 51 unit tests covering all FMJD rules
- Add requirements-dev.txt (pytest, pytest-asyncio)
- Fix scanEngine.ts: parseInt → parseFloat for Scan score, fix comment

+++ frontend/src/lib/scanEngine.ts
@@ -2,7 +2,7 @@ import type { MoveData } from '../types'
 
 export interface ScanInfo {
   depth: number
-  score: number   // centipawns, positive = white advantage
+  score: number   // piece units, positive = white advantage (1.0 ≈ one-piece advantage)
   nodes: number
   nps: number
   pv: string[]    // principal variation in Hub notation
@@ -46,7 +46,7 @@ function parseInfo(msg: string): Partial<ScanInfo> {
   const npsM   = msg.match(/\bnps=(\d+)/)
   const pvM    = msg.match(/\bpv="([^"]*)"/)
   if (depthM) result.depth = parseInt(depthM[1])
-  if (scoreM) result.score = parseInt(scoreM[1])
+  if (scoreM) result.score = parseFloat(scoreM[1])
   if (nodesM) result.nodes = parseInt(nodesM[1])
   if (npsM)   result.nps   = parseInt(npsM[1])
   if (pvM)    result.pv    = pvM[1].trim().split(/\s+/).filter(Boolean)
